requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.3.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
blis==0.7.10
boto3==1.26.142
botocore==1.29.142
Brotli==1.0.9
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clean-text==0.6.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comtypes==1.2.0
confection==0.1.1
cryptography==41.0.3
cymem==2.0.7
decorator==4.4.2
dill==0.3.7
elevenlabs==0.2.17
emoji==1.7.0
exceptiongroup==1.1.3
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.12.3
Fiona==1.9.4.post1
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
fsspec==2023.9.0
ftfy==6.1.1
future==0.18.3
geopandas==0.13.2
greenlet==2.0.2
gTTS==2.3.2
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.8
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
langcodes==3.3.0
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mutagen==1.47.0
mysql-connector==2.2.9
networkx==3.1
numpy==1.25.2
osmnx==1.6.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathos==0.3.1
pathy==0.10.2
pickleshare==0.7.5
Pillow==9.5.0
playwright==1.34.0
pox==0.3.3
ppft==1.7.6.7
praw==7.7.0
prawcore==2.3.0
preshed==3.0.8
proglog==0.1.10
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.12
pyee==9.0.4
PyExecJS==1.5.1
Pygments==2.16.1
pypiwin32==223
pyproj==3.6.0
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.4.1
s3transfer==0.6.2
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.6.2
sympy==1.12
thinc==8.1.12
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.11.8
torch==2.0.1
tqdm==4.66.1
traitlets==5.9.0
transformers==4.29.2
translators==5.7.6
typer==0.7.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.6
update-checker==0.18.0
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.6
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
yt-dlp==2023.7.6
